[UI-86] Combine Form, FormHandler and FormField into single package
