+++ packages/credit-card/README.md
@@ -12,8 +12,7 @@ using `npm install @talixo/credit-card --save` or `yarn add @talixo/credit-card`
 Your package should additionally have some extra dependencies:
 
 - `@talixo/combo-box: ^1.0.0-alpha.5`
-- `@talixo/form-field: ^1.0.0-alpha.5`
-- `@talixo/form-handler: ^1.0.0-alpha.5`
+- `@talixo/form: ^1.0.0-alpha.5`
 - `@talixo/icon: ^1.0.0-alpha.5`
 - `@talixo/shared: ^1.0.0-alpha.5`
 - `@talixo/text-input: ^1.0.0-alpha.5`
