--- packages/form-field/.npmignore
@@ -1,23 +1,0 @@
-# Folder view configuration files
-.DS_Store
-Desktop.ini
-
-# Thumbnail cache files
-._*
-*~
-Thumbs.db
-
-# Files that might appear on external disks
-.Spotlight-V100
-.Trashes
-
-# npm
-npm-debug.log
-
-# IDEs stuff
-.idea
-
-# Development tools
-tests
-rollup.config.js
-stories.js
