--- packages/form-field/index.js
@@ -1,5 +1,0 @@
-import FormField from './src/FormField'
-
-export {
-  FormField
-}
