--- packages/form-field/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "@talixo/form-field",
-  "version": "1.0.0-alpha.5",
-  "description": "UI Component which represents FormField",
-  "main": "dist/index.js",
-  "module": "index.js",
-  "jsnext:main": "index.js",
-  "scripts": {
-    "test": "node ../../scripts/runTests.js --only form-field",
-    "prepublishOnly": "npm run build",
-    "build": "node ../../scripts/buildPackages.js --only form-field"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "author": "Public in Motion GmbH",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/PublicInMotionGmbH/ui-kit.git"
-  },
-  "bugs": {
-    "url": "https://github.com/PublicInMotionGmbH/ui-kit/issues"
-  },
-  "license": "MIT",
-  "devDependencies": {
-    "@talixo/checkbox": "^1.0.0-alpha.5",
-    "@talixo/number-input": "^1.0.0-alpha.5",
-    "@talixo/shared": "^1.0.0-alpha.5",
-    "@talixo/text-input": "^1.0.0-alpha.5",
-    "prop-types": "^15.6.1",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0"
-  },
-  "peerDependencies": {
-    "@talixo/checkbox": "^1.0.0-alpha.5",
-    "@talixo/number-input": "^1.0.0-alpha.5",
-    "@talixo/shared": "^1.0.0-alpha.5",
-    "@talixo/text-input": "^1.0.0-alpha.5",
-    "prop-types": "^15.6.1",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0"
-  }
-}
