--- packages/form-field/styles/config.sass
@@ -1,9 +1,0 @@
-@import '~@talixo/shared/styles/config'
-
-$form-field-label-padding-right: dp(2)
-
-$form-field-hint-padding-left: dp(2)
-
-$form-field-error-color: $x-color-error
-$form-field-hint-color: #666
-$form-field-warning-color: $x-color-warning
