--- packages/form-field/styles/geometry.sass
@@ -1,55 +1,0 @@
-@import './config'
-
-@include module('form-field')
-  display: flex
-  flex-wrap: wrap
-  margin: dp(2) 0
-
-  *
-    box-sizing: border-box
-
-  &__label
-    display: flex
-    align-items: center
-    justify-content: flex-end
-    min-width: 30%
-    max-width: 30%
-    text-align: right
-    font-size: dp(1.75)
-    padding-right: $form-field-label-padding-right
-    font-weight: 600
-
-  &__hint
-    flex: 1 auto
-    display: flex
-    align-items: center
-    padding-left: $form-field-hint-padding-left
-    font-size: dp(1.3)
-    line-height: 1.2em
-    color: $form-field-hint-color
-
-  &__input-wrapper
-    display: flex
-    min-width: 70%
-    max-width: 70%
-
-    > *
-      max-width: dp(36)
-      min-width: dp(36)
-
-  &__error, &__warning
-    min-width: 70%
-    max-width: 70%
-    font-size: 0.75rem
-    font-weight: 600
-    margin-top: 0.5rem
-    line-height: 1em
-
-    &--labeled
-      margin-left: 30%
-
-  &__error
-    color: $form-field-error-color
-
-  &__warning
-    color: $form-field-warning-color
