--- packages/form-field/styles/main.sass
@@ -1,2 +1,0 @@
-@import './geometry'
-@import './theming'
