--- packages/form-field/styles/theming.sass
@@ -1,3 +1,0 @@
-@import './config'
-
-@include module('form-field')
