--- packages/form-handler/index.js
@@ -1,5 +1,0 @@
-import FormHandler from './src/FormHandler'
-
-export {
-  FormHandler
-}
