--- packages/form-handler/package.json
@@ -1,54 +1,0 @@
-{
-  "name": "@talixo/form-handler",
-  "version": "1.0.0-alpha.5",
-  "description": "UI Component which represents Form Handler",
-  "main": "dist/index.js",
-  "module": "index.js",
-  "jsnext:main": "index.js",
-  "scripts": {
-    "test": "node ../../scripts/runTests.js --only form-handler",
-    "prepublishOnly": "npm run build",
-    "build": "node ../../scripts/buildPackages.js --only form-handler"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "author": "Public in Motion GmbH",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/PublicInMotionGmbH/ui-kit.git"
-  },
-  "bugs": {
-    "url": "https://github.com/PublicInMotionGmbH/ui-kit/issues"
-  },
-  "license": "MIT",
-  "devDependencies": {
-    "@talixo/button": "^1.0.0-alpha.5",
-    "@talixo/calendar": "^1.0.0-alpha.5",
-    "@talixo/checkbox": "^1.0.0-alpha.5",
-    "@talixo/combo-box": "^1.0.0-alpha.5",
-    "@talixo/control-group": "^1.0.0-alpha.5",
-    "@talixo/fieldset": "^1.0.0-alpha.5",
-    "@talixo/form-field": "^1.0.0-alpha.5",
-    "@talixo/icon": "^1.0.0-alpha.5",
-    "@talixo/number-input": "^1.0.0-alpha.5",
-    "@talixo/phone-input": "^1.0.0-alpha.5",
-    "@talixo/radio-input": "^1.0.0-alpha.5",
-    "@talixo/shared": "^1.0.0-alpha.5",
-    "@talixo/switcher": "^1.0.0-alpha.5",
-    "@talixo/text-input": "^1.0.0-alpha.5",
-    "formik": "^0.11.11",
-    "prop-types": "^15.6.1",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0",
-    "yup": "^0.24.1"
-  },
-  "peerDependencies": {
-    "@talixo/form-field": "^1.0.0-alpha.5",
-    "@talixo/shared": "^1.0.0-alpha.5",
-    "formik": "^0.11.11",
-    "prop-types": "^15.6.1",
-    "react": "^16.2.0",
-    "react-dom": "^16.2.0"
-  }
-}
