--- packages/form-handler/styles/config.sass
@@ -1,1 +1,0 @@
-@import '~@talixo/shared/styles/config'
