+++ packages/form/index.js
@@ -1,5 +1,9 @@
 import Form from './src/Form'
+import FormHandler from './src/FormHandler'
+import FormField from './src/FormField'
 
 export {
-  Form
+  Form,
+  FormHandler,
+  FormField
 }
