+++ packages/form/package.json
@@ -29,21 +29,24 @@
     "@talixo/combo-box": "^1.0.0-alpha.5",
     "@talixo/control-group": "^1.0.0-alpha.5",
     "@talixo/fieldset": "^1.0.0-alpha.5",
-    "@talixo/form-field": "^1.0.0-alpha.5",
     "@talixo/icon": "^1.0.0-alpha.5",
     "@talixo/number-input": "^1.0.0-alpha.5",
     "@talixo/phone-input": "^1.0.0-alpha.5",
     "@talixo/shared": "^1.0.0-alpha.5",
     "@talixo/switcher": "^1.0.0-alpha.5",
     "@talixo/text-input": "^1.0.0-alpha.5",
+    "formik": "^0.11.11",
     "prop-types": "^15.6.1",
     "react": "^16.2.0",
-    "react-dom": "^16.2.0"
+    "react-dom": "^16.2.0",
+    "yup": "^0.24.1"
   },
   "peerDependencies": {
     "@talixo/shared": "^1.0.0-alpha.5",
+    "formik": "^0.11.11",
     "prop-types": "^15.6.1",
     "react": "^16.2.0",
-    "react-dom": "^16.2.0"
+    "react-dom": "^16.2.0",
+    "yup": "^0.24.1"
   }
 }
