+++ packages/form/styles/config.sass
@@ -1,3 +1,10 @@
 @import '~@talixo/shared/styles/config'
 
 $color-border: #662FB3
+
+
+$form-field-label-padding-right: dp(2)
+$form-field-hint-padding-left: dp(2)
+$form-field-error-color: $x-color-error
+$form-field-hint-color: #666
+$form-field-warning-color: $x-color-warning
